requirements.txt
asgiref==3.8.1
Django==5.1.1
psycopg2-binary==2.9.9
djangorestframework==3.15.2
django-bootstrap5==24.3
gunicorn==23.0.0
sqlparse==0.5.1
typing-extensions==4.12.2
tzdata==2024.1
dj-database-url==2.2.0
django-on-heroku==1.1.2
whitenoise==6.7.0
packaging==24.1
python-decouple==3.8
django-bootstrap5==24.3
boto3==1.26.59
django-storages==1.13.2
django-allauth==65.0.2
requests==2.32.3
PyJWT==2.9.0
cryptography==43.0.3
python-dotenv==1.0.1







